Add rdb/gorm support for GetVendorProducts and GetCpesByVendorProduct
This adds in rdb/gorm support for GetVendorProducts and GetCpesByVendorProduct.

I refactored the tests, so they could be shared by rdb and redis. In doing so, I noticed a couple changes I needed to make to the expected strings.
1. I’m seeing a `\\` inserted by go-cpe
2. The cpes returned by TestGetCpesByVendorProduct were different.

I’m most interested in adding some fuzzy searching here. So I used `LIKE`